backend/requirements.txt
django
pytz
uuid
djangorestframework
markdown
django-filter
drf-yasg
django-rest-swagger